feat: GET_INFO to element.sh
Added GET_INFO to element.sh so that the case of not being able
to find an element resulted in the proper output. If the
database seach returned an empty result the script will say
as much. And to avoid having multiple instances of calling
GET_INFO for each case, it was moved to the end of the if cases
and wont return anything if given an empty result, in the case
of no database matches.